Revert "Convert testing/test-util.js to typescript (#2243)"
This reverts commit 104ff0e0a11cb8d4591cf55c1b7d3ea1b2e948b3.
