+++ runtime/test/arc-tests.js
@@ -11,7 +11,7 @@
 import {Arc} from '../ts-build/arc.js';
 import {assert} from './chai-web.js';
 import {SlotComposer} from '../ts-build/slot-composer.js';
-import * as util from '../ts-build/testing/test-util.js';
+import * as util from '../testing/test-util.js';
 import {handleFor} from '../ts-build/handle.js';
 import {Manifest} from '../ts-build/manifest.js';
 import {Loader} from '../ts-build/loader.js';
