+++ runtime/test/debug/outer-port-attachments-tests.js
@@ -9,7 +9,7 @@
  */
 
 import {assert} from '../chai-web.js';
-import * as util from '../../ts-build/testing/test-util.js';
+import * as util from '../../testing/test-util.js';
 import {StubLoader} from '../../testing/stub-loader.js';
 import {DevtoolsForTests} from '../../debug/devtools-connection.js';
 import {Random} from '../../ts-build/random.js';
